requirements.txt
altair                    4.2.2       
asttokens                 2.4.1       
attrs                     23.2.0      
avro                      1.11.0      
boto3                     1.34.53     
botocore                  1.34.53
certifi                   2024.2.2
cffi                      1.16.0
charset-normalizer        3.3.2
click                     8.1.7
colorama                  0.4.6
comm                      0.2.1
confluent-kafka           2.3.0
contourpy                 1.2.0
cryptography              42.0.5
cycler                    0.12.1
dataclasses               0.6
debugpy                   1.8.1
decorator                 5.1.1
docker                    7.0.0
entrypoints               0.4
exceptiongroup            1.2.0
executing                 2.0.1
fastavro                  1.8.2
fastjsonschema            2.19.1
feature-engine            1.6.2
fonttools                 4.49.0
fsspec                    2024.2.0
furl                      2.1.3
future                    1.0.0
great-expectations        0.14.13
greenlet                  3.0.3
hopsworks                 3.4.4
hsfs                      3.4.7
hsml                      3.4.6
idna                      3.6
importlib-metadata        7.0.1
ipykernel                 6.29.3
ipython                   8.22.1
javaobj-py3               0.4.3
jedi                      0.19.1
Jinja2                    3.0.3
jmespath                  1.0.1
joblib                    1.3.2
jsonpatch                 1.33
jsonpointer               2.4
jsonschema                4.21.1
jsonschema-specifications 2023.12.1
jupyter_client            8.6.0
jupyter_core              5.7.1
kiwisolver                1.4.5
MarkupSafe                2.0.1
matplotlib                3.8.3
matplotlib-inline         0.1.6
mistune                   3.0.2
mock                      5.1.0
nbformat                  5.9.2
nest-asyncio              1.6.0
numpy                     1.26.4
orderedmultidict          1.0.1
packaging                 23.2
pandas                    2.1.4
parso                     0.8.3
patsy                     0.5.6
pillow                    10.2.0
pip                       24.0
platformdirs              4.2.0
prompt-toolkit            3.0.43
psutil                    5.9.8
pure-eval                 0.2.2
pyarrow                   15.0.0
pyasn1                    0.5.1
pyasn1-modules            0.3.0
pycparser                 2.21
pycryptodomex             3.20.0
Pygments                  2.17.2
PyHopsHive                0.6.4.1.dev0
pyhumps                   1.6.1
pyjks                     20.0.0
PyMySQL                   1.1.0
pyparsing                 2.4.7
python-dateutil           2.9.0
python-dotenv             1.0.1
pytz                      2024.1
pywin32                   306
pyzmq                     25.1.2
referencing               0.33.0
requests                  2.31.0
rpds-py                   0.18.0
ruamel.yaml               0.17.17
s3transfer                0.10.0
scikit-learn              1.4.1.post1
scipy                     1.12.0
setuptools                68.2.0
six                       1.16.0
SQLAlchemy                2.0.27
stack-data                0.6.3
statsmodels               0.14.1
termcolor                 2.4.0
threadpoolctl             3.3.0
thrift                    0.16.0
toolz                     0.12.1
tornado                   6.4
tqdm                      4.66.2
traitlets                 5.14.1
twofish                   0.3.0
typing_extensions         4.10.0
tzdata                    2024.1
tzlocal==5.2
urllib3==1.26.18
wcwidth==0.2.13
wheel==0.41.2
xgboost==2.0.3
zipp==3.17.0